fix: allow for dots in KV Store name
Fastly allows for special characters (like `.`) in the KV store name,
but since the configuration is in toml, these characters need to be
either escaped, or wrapped in quotes to be treated as a single string
rather than configuration.

This change wraps the kvStoreName variable in double quotes to prevent
toml interpolating the name.

Example error:

```
Don't do this!

node:internal/process/esm_loader:40
      internalBinding('errors').triggerUncaughtException(
                                ^
No
(Use `node --trace-uncaught ...` to show where the exception was thrown)
```

Signed-off-by: Neil Hanlon <neil@shrug.pw>